Add GameAdministrator update package.json
This change creates the GameAdministrator which is
responsible for running the game. The admin shows the board
and current available moves, gets the current player's move,
and updates the board. This is now used in favor of the
DebugDrawer which has been removed.

This also updates the package.json to include necessary
npm start command and basic info.