Add validation that no check happens on king move
This change updates the moves possible for the king to ensure
that the king of the same color (i.e. the current piece) isn't in check
by performing the current move. This simulates the move on the board
and then verifies that the king isn't in check.

This can be easily added to the other pieces to validate that no illegal moves are made.