Add pawn check for if king is in check
This change updates the pawn to check whether or not its own
king is in check before allowing a move. Added a test as well.

TODO: figure out mocking because this currently relies on the rook
class to function properly for the test to pass.

Also, I don't understand Node. Probably need to give Michael a ring
(or post on StackOverflow).