Simplify getMovablePositions
To get movable positions for pieces that slide, after finding all watched positions,
then the positions that are unoccupied by allies and that do not cause a check on
the friendly king can be filtered out. This change simplifies the code to perform
the filters and applies them to the rook, king, and bishop.